refactor: move logic that changes amount raised output to a new class
