+++ src/DonationForms/ServiceProvider.php
@@ -8,9 +8,9 @@ use Give\DonationForms\Actions\DispatchDonateControllerSubscriptionCreatedListen
 use Give\DonationForms\Actions\SanitizeDonationFormPreviewRequest;
 use Give\DonationForms\Actions\StoreBackwardsCompatibleFormMeta;
 use Give\DonationForms\AsyncData\Actions\GetAsyncFormDataForListView;
+use Give\DonationForms\AsyncData\Actions\GiveGoalProgressStats;
 use Give\DonationForms\AsyncData\Actions\LoadAsyncDataAssets;
 use Give\DonationForms\AsyncData\AdminFormListView\AdminFormListView;
-use Give\DonationForms\AsyncData\AsyncDataHelpers;
 use Give\DonationForms\AsyncData\FormGrid\FormGridView;
 use Give\DonationForms\Blocks\DonationFormBlock\Block as DonationFormBlock;
 use Give\DonationForms\Controllers\DonationConfirmationReceiptViewController;
@@ -101,7 +101,8 @@ class ServiceProvider implements ServiceProviderInterface
         Hooks::addAction('wp_ajax_nopriv_givewp_get_form_async_data_for_list_view', GetAsyncFormDataForListView::class);
 
         // Filter from give_goal_progress_stats() function which is used by the admin form list views and form grid view
-        Hooks::addFilter('give_goal_amount_raised_output', AsyncDataHelpers::class, 'maybeChangeAmountRaisedOutput', 1,
+        Hooks::addFilter('give_goal_amount_raised_output', GiveGoalProgressStats::class,
+            'maybeChangeAmountRaisedOutput', 1,
             2);
 
         // Form Grid
